Updating `numHexagons` with its closed-form (#409)
* updating `numHexagons` with its closed-form

* Sure Mr. Formatter, you're the boss

* move math notes to `h3api.h.in`

* add test for `numHexagons`

* remove `generateNumHexagons.c`

* remove `generateNumHexagons.c` from `CMakeLists.txt`

* bug in `maxH3ToChildrenSize`: `int` output should probably be `uint64_t`

* change `_ipow` and `maxH3ToChildrenSize` to use `uint64_t` inputs/outputs to fix int overflow bug

* ok, formatting, ok

* changing things back to uint64_t for consistency

* drop the `unsigned` stuff that I'd added

* have `_ipow` output `int64_t` instead of `uint64_t` to potentially avoid
overflow/wrap-around issues if subtraction is performed on outputs

* updating changelog

* BROKEN: just to show that the `H3ToChildren` function needs the full
7^n working space, and breaks with the smaller allocation for pentagons.

* revert breaking change

* change `_ipow` inputs/output back to `int64_t`

+++ src/h3lib/include/mathExtensions.h
@@ -20,12 +20,14 @@
 #ifndef MATHEXTENSIONS_H
 #define MATHEXTENSIONS_H
 
+#include <stdint.h>
+
 /**
  * MAX returns the maximum of two values.
  */
 #define MAX(a, b) (((a) > (b)) ? (a) : (b))
 
 // Internal functions
-int _ipow(int base, int exp);
+int64_t _ipow(int64_t base, int64_t exp);
 
 #endif
